Update README.md
Hi,

A while back I wrote two blog posts about writing an SDF generator using Compute Shaders in Unity. It covers the code and the maths behind it. The output is stored in 3d textures and I also show examples of how to visualize the SDF data. The blog posts are accompanied by a GitHub repo with the implementation.